+++ packages/components-js/tests/vrt/specs/switch.vrt.ts
@@ -38,6 +38,7 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
       );
 
       await forceHoverState(page, '.hover > p-switch >>> button');
+      await forceFocusState(page, '.focus > p-switch'); // native outline should not be visible
       await forceFocusState(page, '.focus > p-switch >>> button');
       await forceFocusHoverState(page, '.focus-hover > p-switch >>> button');
     })