+++ packages/components/src/components/action/switch/switch.tsx
@@ -4,7 +4,6 @@ import {
   attachComponentCss,
   getPrefixedTagNames,
   improveButtonHandlingForCustomElement,
-  improveFocusHandlingForCustomElement,
   isDisabledOrLoading,
 } from '../../../utils';
 import { getComponentCss } from './switch-styles';
@@ -13,7 +12,7 @@ export type SwitchChangeEvent = { checked: boolean };
 
 @Component({
   tag: 'p-switch',
-  shadow: true,
+  shadow: { delegatesFocus: true },
 })
 export class Switch {
   @Element() public host!: HTMLElement;
@@ -36,7 +35,9 @@ export class Switch {
   /** Disables the switch and shows a loading indicator. No events will be triggered while loading state is active. */
   @Prop() public loading?: boolean = false;
 
-  /** To remove the element from tab order. */
+  /** To remove the element from tab order.
+   * @deprecated since v2.8.0, use tabindex="-1" instead
+   */
   @Prop() public tabbable?: boolean = true;
 
   /** Adapts the switch color depending on the theme. */
@@ -53,7 +54,6 @@ export class Switch {
   }
 
   public componentDidLoad(): void {
-    improveFocusHandlingForCustomElement(this.host);
     improveButtonHandlingForCustomElement(
       this.host,
       () => 'button',